src/circuits/qubit_routing.py
# src/circuits/qubit_routing.py

from qiskit import QuantumCircuit, transpile
from qiskit.compiler import transpile
from typing import Dict, Any

class QubitRouter:
    """
    Implementation of qubit routing and allocation algorithms for quantum circuits.
    
    This class handles the decomposition of quantum circuits to match specific
    hardware constraints and optimize qubit allocation.
    """
    
    def __init__(self, basis_gates: list = None):
        """
        Initialize the qubit router with specified basis gates.
        
        Args:
            basis_gates: List of basis gates to use for decomposition
        """
        self.basis_gates = basis_gates or ['cx', 'u']
    
    def decompose_circuit(self, circuit: QuantumCircuit, optimization_level: int = 1) -> QuantumCircuit:
        """
        Decompose a quantum circuit into the specified basis gates.
        
        Args:
            circuit: Input quantum circuit to decompose
            optimization_level: Optimization level for transpilation (0-3)
            
        Returns:
            QuantumCircuit: Decomposed circuit
        """
        return transpile(circuit, 
                        basis_gates=self.basis_gates,
                        optimization_level=optimization_level)
    
    def optimize_layout(self, circuit: QuantumCircuit, coupling_map: list) -> Dict[str, Any]:
        """
        Optimize qubit layout for a given coupling map using SABRE algorithm.
        
        Args:
            circuit: Input quantum circuit
            coupling_map: Hardware coupling map constraints
            
        Returns:
            Dict containing optimized circuit and metrics
        """
        optimized = transpile(circuit,
                            basis_gates=self.basis_gates,
                            coupling_map=coupling_map,
                            layout_method='sabre',
                            routing_method='sabre')
        
        return {
            'circuit': optimized,
            'depth': optimized.depth(),
            'size': optimized.size(),
            'gate_counts': optimized.count_ops()
        }